Sharing dependencies between microfrontends
